Add final_touches to _special_treatment
